+++ pandas/core/internals.py
@@ -25,6 +25,7 @@ from pandas.core.indexing import maybe_convert_indices, length_of_indexer
 from pandas.core.categorical import Categorical, maybe_to_categorical
 from pandas.tseries.index import DatetimeIndex
 import pandas.core.common as com
+import pandas.core.missing as mis
 import pandas.core.convert as convert
 from pandas.sparse.array import _maybe_to_sparse, SparseArray
 import pandas.lib as lib
@@ -852,7 +853,7 @@ class Block(PandasObject):
 
         # a fill na type method
         try:
-            m = com._clean_fill_method(method)
+            m = mis._clean_fill_method(method)
         except:
             m = None
 
@@ -870,7 +871,7 @@ class Block(PandasObject):
                                                mgr=mgr)
         # try an interp method
         try:
-            m = com._clean_interp_method(method, **kwargs)
+            m = mis._clean_interp_method(method, **kwargs)
         except:
             m = None
 
@@ -909,7 +910,7 @@ class Block(PandasObject):
         values = self.values if inplace else self.values.copy()
         values, _, fill_value, _ = self._try_coerce_args(values, fill_value)
         values = self._try_operate(values)
-        values = com.interpolate_2d(values,
+        values = mis.interpolate_2d(values,
                                     method=method,
                                     axis=axis,
                                     limit=limit,
@@ -949,8 +950,8 @@ class Block(PandasObject):
 
             # process a 1-d slice, returning it
             # should the axis argument be handled below in apply_along_axis?
-            # i.e. not an arg to com.interpolate_1d
-            return com.interpolate_1d(index, x, method=method, limit=limit,
+            # i.e. not an arg to mis.interpolate_1d
+            return mis.interpolate_1d(index, x, method=method, limit=limit,
                                       limit_direction=limit_direction,
                                       fill_value=fill_value,
                                       bounds_error=False, **kwargs)
@@ -2357,7 +2358,7 @@ class SparseBlock(NonConsolidatableMixIn, Block):
     def interpolate(self, method='pad', axis=0, inplace=False,
                     limit=None, fill_value=None, **kwargs):
 
-        values = com.interpolate_2d(
+        values = mis.interpolate_2d(
             self.values.to_dense(), method, axis, limit, fill_value)
         return self.make_block_same_class(values=values,
                                           placement=self.mgr_locs)
@@ -3773,7 +3774,7 @@ class SingleBlockManager(BlockManager):
 
         # fill if needed
         if method is not None or limit is not None:
-            new_values = com.interpolate_2d(new_values, method=method,
+            new_values = mis.interpolate_2d(new_values, method=method,
                                             limit=limit, fill_value=fill_value)
 
         if self._block.is_sparse:
